# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bengkol Workshop & Repair-Shop Management API
  description: |
    Production-ready REST API for the Bengkol auto repair-shop platform.
    Connects vehicle owners with nearby repair shops for slot booking, real-time queues, service catalogs, inventory, and reviews.
  version: 1.0.0
  contact:
    name: Bengkol Engineering
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api-dev.bengkol.com
    description: Development (Develop) environment
  - url: https://api-uat.bengkol.com
    description: User Acceptance Testing (UAT) environment
  - url: https://api-staging.bengkol.com
    description: Staging environment
  - url: https://api.bengkol.com
    description: Production (Prod) environment
paths:
...
  /api/v1/me:
    get:
      summary: Get authenticated user profile
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
  /api/v1/workshops:
    get:
      summary: List workshops
      tags:
        - Workshops
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: List of workshops
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopListResponse'
    post:
      summary: Create a new workshop
      tags:
        - Workshops
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateWorkshopRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkshopRequest'
      responses:
        '201':
          description: Workshop created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
  /api/v1/workshops/nearby:
    get:
      summary: Find nearby workshops via PostGIS spatial search
      tags:
        - Workshops
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: lng
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: radius
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Nearby workshops ordered by distance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyWorkshopsResponse'
  /api/v1/workshops/{id}:
    get:
      summary: Get workshop detail
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workshop details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
    patch:
      summary: Update workshop profile
      tags:
        - Workshops
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateWorkshopRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkshopRequest'
      responses:
        '200':
          description: Updated workshop details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
  /api/v1/workshops/{id}/photos/{photoID}:
    get:
      summary: Retrieve workshop photo binary data
      description: Returns the raw binary image stream stored directly in the database.
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: photoID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Binary photo image stream
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workshops/{id}/operating-hours:
    get:
      summary: Get workshop operating hours
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of operating schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatingHoursResponse'
    put:
      summary: Set weekly operating hours
      tags:
        - Workshops
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OperatingHourInput'
      responses:
        '200':
          description: Updated operating hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatingHoursResponse'
  /api/v1/workshops/{id}/employees:
    get:
      summary: List employees for a workshop
      description: Returns a paginated list of employees for the specified workshop. Accessible by owner, admin, or active workshop employees.
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: List of workshop employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workshop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add a new employee to a workshop
      description: Adds an employee to the workshop. Accessible by workshop owner, manager, or system admin. Managers cannot assign the OWNER role.
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/v1/workshops/{id}/employees/me:
    get:
      summary: Get caller's employee profile and permissions in workshop
      description: Returns the authenticated user's active employee role and calculated permissions for the workshop.
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current employee profile and permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User is not an employee of this workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workshops/{id}/employees/{employeeId}:
    get:
      summary: Get workshop employee by ID
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: employeeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update workshop employee
      description: Updates employee role, contact info, specialization, status, or notes. Managers cannot modify workshop owners.
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: employeeId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
    delete:
      summary: Delete or remove employee from workshop
      description: Removes employee from workshop. Managers cannot delete workshop owners. Employees cannot delete themselves.
      tags:
        - Employees
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: employeeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cannot delete own employee record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workshops/{id}/available-slots:
    get:
      summary: Get available booking slots for a workshop on a date
      description: Returns capacity-checked available time slots within operating hours for the given date.
      tags:
        - Bookings
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: date
          schema:
            type: string
            format: date
          example: "2026-09-28"
      responses:
        '200':
          description: Available slots schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableSlotsResponse'
  /api/v1/bookings:
    post:
      summary: Book a service appointment slot (with optional spare parts)
      description: |
        Creates a concurrency-safe service appointment slot reservation with optional spare parts pre-order.
        Double-booking is strictly prevented via transactional locking.
        If spare parts are included, stock availability is verified and deducted atomically from workshop inventory.
      tags:
        - Bookings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            examples:
              with_spare_parts:
                summary: Booking with service and spare parts
                value:
                  workshop_id: "00000000-0000-0000-0000-000000000001"
                  service_id: "00000000-0000-0000-0000-000000000002"
                  booking_date: "2026-09-28"
                  booking_time: "10:00:00"
                  customer_notes: "Ganti oli mesin dan busi"
                  spare_parts:
                    - spare_part_id: "00000000-0000-0000-0000-000000000003"
                      quantity: 2
                    - spare_part_id: "00000000-0000-0000-0000-000000000004"
                      quantity: 1
              service_only:
                summary: Booking service only
                value:
                  workshop_id: "00000000-0000-0000-0000-000000000001"
                  service_id: "00000000-0000-0000-0000-000000000002"
                  booking_date: "2026-09-28"
                  booking_time: "10:00:00"
                  customer_notes: "Servis rutin berkala"
      responses:
        '201':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          description: Workshop closed or invalid time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slot capacity exhausted (BOOKING_SLOT_UNAVAILABLE) or insufficient spare part stock (INSUFFICIENT_STOCK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Input validation failed (inactive service/spare part, workshop mismatch, past date)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/v1/bookings/{id}:
    get:
      summary: Get booking details
      description: Returns booking details including reserved slot, booked service, and any booked spare parts.
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
  /api/v1/bookings/{id}/cancel:
    post:
      summary: Cancel a booking
      description: |
        Cancels an existing reservation, restores slot capacity, and automatically restores any reserved spare parts stock back to workshop inventory. Allowed for the customer, workshop owner, or admin.
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking cancelled and reserved stock restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cannot cancel a booking that is completed or in service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not booking customer, workshop owner, or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/bookings:
    get:
      summary: List authenticated customer's bookings
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: List of customer bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
  /api/v1/workshops/{id}/bookings:
    get:
      summary: List bookings for a workshop (Owner / Admin)
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Workshop bookings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
...
  /api/v1/me/vehicles:
    get:
      summary: List authenticated user's vehicles (convenience endpoint)
      tags:
        - Vehicles
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of customer's vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
...
  /api/v1/bookings/{id}/check-in:
    post:
      summary: Check in for a confirmed booking on service day to get a queue ticket
      tags:
        - Queue
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Queue ticket issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueDetailResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkshopListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workshop'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    NearbyWorkshopsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workshop'
    WorkshopDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Workshop'
...
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
...